Simplify logic somewhat
While reviewing this file we felt that the logic was a bit convoluted.
Since if it's a promise it cannot be a FSA, that check is unnecessary
for the thunk support.
Additionally, the main meat of the function (happy path) should be the
least indented, so move the isPromise check to an early return.

Reviewed-by: Piotr Anielski <piotr.anielski@smarkets.com>